Users entity created; Security partially done; Storage service almost done

+++ Backend/src/main/java/com/backend/Backend/repositories/SmestajRepo.java
@@ -3,5 +3,5 @@ package com.backend.Backend.repositories;
 import com.backend.Backend.dataTypes.Smestaj;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SmestajRepo extends CrudRepository<Smestaj, Integer> {
+public interface SmestajRepo extends CrudRepository<Smestaj, Long> {
 }
